Share logic between step classes
